Fix unsigned comparison problem for conditional branch instructions

+++ test/tools/reference_simulator/asm_reference_simulator.cpp
@@ -392,7 +392,7 @@ void simulateMIPSHelper(unordered_map<uint32_t, uint32_t> &memory, uint32_t pc, 
             tie(ignore, sReg, immediate) = decodeImmediateType(instruction);
             pc++;
             simulateMIPSHelper(memory, pc, regs, lo, hi, memInstructionStartIdx, true); // execute branch delay slot instruction
-            if (regs[sReg] > 0)
+            if (static_cast<int16_t>(regs[sReg]) > 0)
             {
                 // Use signed immediates as relative branches could be negative
                 pc += (static_cast<int16_t>(immediate) << 2) / 4;
@@ -409,7 +409,7 @@ void simulateMIPSHelper(unordered_map<uint32_t, uint32_t> &memory, uint32_t pc, 
             tie(ignore, sReg, immediate) = decodeImmediateType(instruction);
             pc++;
             simulateMIPSHelper(memory, pc, regs, lo, hi, memInstructionStartIdx, true); // execute branch delay slot instruction
-            if (regs[sReg] <= 0)
+            if (static_cast<int16_t>(regs[sReg]) <= 0)
             {
                 // Use signed immediates as relative branches could be negative
                 pc += (static_cast<int16_t>(immediate) << 2) / 4;
@@ -431,7 +431,7 @@ void simulateMIPSHelper(unordered_map<uint32_t, uint32_t> &memory, uint32_t pc, 
             {
             case 0b00000: // BLTZ
             {
-                if (regs[sReg] < 0)
+                if (static_cast<int16_t>(regs[sReg]) < 0)
                 {
                     // Use signed immediates as relative branches could be negative
                     pc += (static_cast<int16_t>(immediate) << 2) / 4;
@@ -444,7 +444,7 @@ void simulateMIPSHelper(unordered_map<uint32_t, uint32_t> &memory, uint32_t pc, 
             }
             case 0b00001: // BGEZ
             {
-                if (regs[sReg] >= 0)
+                if (static_cast<int16_t>(regs[sReg]) >= 0)
                 {
                     // Use signed immediates as relative branches could be negative
                     pc += (static_cast<int16_t>(immediate) << 2) / 4;
@@ -458,7 +458,7 @@ void simulateMIPSHelper(unordered_map<uint32_t, uint32_t> &memory, uint32_t pc, 
             case 0b10001: // BGEZAL
             {
 
-                if (regs[sReg] >= 0)
+                if (static_cast<int16_t>(regs[sReg]) >= 0)
                 {
                     // Use signed immediates as relative branches could be negative
                     pc += (static_cast<int16_t>(immediate) << 2) / 4;
@@ -473,7 +473,7 @@ void simulateMIPSHelper(unordered_map<uint32_t, uint32_t> &memory, uint32_t pc, 
             }
             case 0b10000: // BLTZAL
             {
-                if (regs[sReg] < 0)
+                if (static_cast<int16_t>(regs[sReg]) < 0)
                 {
                     // Use signed immediates as relative branches could be negative
                     pc += (static_cast<int16_t>(immediate) << 2) / 4;
